% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gene_sets_lfc_grna.R
\name{annotate_gene_sets_lfc_grna}
\alias{annotate_gene_sets_lfc_grna}
\title{annotate_gene_sets_lfc_grna}
\usage{
annotate_gene_sets_lfc_grna(qc)
}
\arguments{
\item{qc}{An object of class `fgcQC` (`fgcQC::QC_fgc_crispr_data`).}
}
\value{
An object of class `lfc_gene_sets` which contains a data frame with a `type` column referring to control/treatment and a `gene_set` column.
}
\description{
Annotates gene sets in log2FC gRNA data and combines control and treatment data for plotting.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
