% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_library_width.R
\name{calculate_library_width}
\alias{calculate_library_width}
\title{calculate_library_width}
\usage{
calculate_library_width(counts)
}
\arguments{
\item{counts}{A vector of integer counts.}
}
\value{
A numeric value.
}
\description{
Calculates the library width of a plasmid sample, as defined in Imkeller et al. (2020) - 90th percentile divided by the 10th percentile.
}
\references{
\itemize{
\item Imkeller, K. et al. 2020. gscreend: modelling asymmetric count ratios in CRISPR screens to decrease experiment size and improve phenotype detection. \emph{Genome Biol} \strong{21}, 53.
}
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
