% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sarles_bimodality.R
\name{calculate_sarles_bimodality}
\alias{calculate_sarles_bimodality}
\title{calculate_sarles_bimodality}
\usage{
calculate_sarles_bimodality(data)
}
\arguments{
\item{data}{A vector of integer counts.}
}
\value{
A numeric value.
}
\description{
Calculates Sarle's bimodality metric, which provides a measure of a skewness when the given distribution is unimodal.
}
