% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crispr_libs}
\alias{crispr_libs}
\title{crispr_libs}
\description{
CRISPR (n and a) library files in AZ-CRUK pipeline cleanr format.
}
\details{
A named list of library data frames.
}
\keyword{data}
