% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crispr_libs_annot_data}
\alias{crispr_libs_annot_data}
\title{crispr_libs_annot_data}
\description{
CRISPR library annotation information, e.g. lists of multi-mapping guides.
}
\details{
A named list of library annotation data.
}
\keyword{data}
