% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_b2f_json.R
\name{extract_b2f_json}
\alias{extract_b2f_json}
\title{extract_b2f_json}
\usage{
extract_b2f_json(path)
}
\arguments{
\item{path}{A character string giving a path to the 'Stats.json' output from 'bcl2fastq2'.}
}
\value{
A data frame
}
\description{
Extracts flowcell-level QC metrics from the 'Stats.json' output from 'bcl2fastq2'.
}
