% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PRROC.R
\name{get_PRROC}
\alias{get_PRROC}
\title{get_PRROC}
\usage{
get_PRROC(data, score_col, group_col = NULL, return = "curve")
}
\arguments{
\item{data}{A data frame containing a 'TP' column.}

\item{score_col}{A character string naming a column containing a score metric.}

\item{group_col}{A character string naming the group of 'data'. If `NULL` then no group, defaults to `NULL`.}

\item{return}{A character string indicating what should be returned: if `curve` then the PR curve is added to the data frame, if `auc` then the AUCPrRc value is returned.}
}
\value{
A data frame or a single numeric value.
}
\description{
Returns precision and recall values for a data frame containing TP and a score metric (e.g. Bayes factor), or the AUPrRc.
}
