% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_LOD.R
\name{global_LOD}
\alias{global_LOD}
\title{global_LOD}
\usage{
global_LOD(bthin_out, type, ...)
}
\arguments{
\item{bthin_out}{The output from `seqgendiff`.}

\item{type}{One of `mageck` or `drugz`.}

\item{...}{A set of data}
}
\value{
A numeric value (a log2 fold change).
}
\description{
Find a global Limit Of Detection (LOD), in terms of a log2 fold change, for positive calls produced on data generated by `seqgendiff`.
}
