% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_desirability_scores.R
\name{join_desirability_scores}
\alias{join_desirability_scores}
\title{join_desirability_scores}
\usage{
join_desirability_scores(..., names, type)
}
\arguments{
\item{...}{Two or more objects returned by `crispRutils::read_desirability_scores`.}

\item{names}{A character vector providing names for the desirability data frames in `...` - must be the same length as `...`.}

\item{type}{One of either `neg` or `pos`, indicating negative desirability scores or positive ones respectively.}
}
\value{
An object of class `desirability`.
}
\description{
Joins desirability scores (from the AZ-CRUK CRISPR analysis pipeline) across 2 or more screens.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
