% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_AUPrRc.R
\name{make_AUPrRc_df}
\alias{make_AUPrRc_df}
\title{make_AUPrRc_df}
\usage{
make_AUPrRc_df(data)
}
\arguments{
\item{data}{A data frame containing the following columns: `Recall`, `Precision`, and `gene_set`.}
}
\value{
A data frame of AUPrRcs.
}
\description{
Makes an AUPrRc data frame.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
