% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_AUROC_df.R
\name{make_AUROC_df}
\alias{make_AUROC_df}
\title{make_AUROC_df}
\usage{
make_AUROC_df(data)
}
\arguments{
\item{data}{A data frame containing the following columns: `True_Positive_Rate`, `False_Positive_Rate`, and `gene_set`.}
}
\value{
A data frame of AUROCs.
}
\description{
Makes an AUROC data frame.
}
