% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pipeline_cleanr_library_file.R
\name{make_pipeline_cleanr_library_file}
\alias{make_pipeline_cleanr_library_file}
\title{make_pipeline_cleanr_library_file}
\usage{
make_pipeline_cleanr_library_file(
  file,
  grna_id_column,
  seq_column,
  gene_column,
  csv = FALSE
)
}
\arguments{
\item{file}{A character string naming a path to an sgRNA sequence file.}

\item{grna_id_column}{A character string naming the sgRNA column.}

\item{seq_column}{A character string naming the sequence column.}

\item{gene_column}{A character string naming the sequence column.}

\item{csv}{Logical indicating whether input is a `csv` file or not. Defaults to `FALSE`.}
}
\value{
A data frame.
}
\description{
Prepares a 'cleanr' library file (to be compatible with the AZ-CRUK CRISPR pipeline) from either an sgRNA sequence file in which genomic co-ordinates are encoded in sgRNA IDs, or a csv file format used for TechDev benchmarking libraries.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
