% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lfc_gene_sets.R
\name{plot.lfc_gene_sets}
\alias{plot.lfc_gene_sets}
\title{plot.lfc_gene_sets}
\usage{
\method{plot}{lfc_gene_sets}(x, notch = TRUE, ...)
}
\arguments{
\item{x}{An object of class `lfc_gene_sets` (`crispRutils::annotate_gene_sets_lfc_grna`).}

\item{notch}{Add notches to the boxplot. Defaults to `TRUE`.}

\item{...}{Other arguments to pass to `plot`.}
}
\value{
Plot.
}
\description{
Plot logFC boxplots for gene sets.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
