% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.synthetic_lethal.R
\name{plot.synthetic_lethal}
\alias{plot.synthetic_lethal}
\title{plot.synthetic_lethal}
\usage{
\method{plot}{synthetic_lethal}(
  x,
  type,
  remove_multimappers,
  library_name = NULL,
  library_type = NULL,
  library_annotation_version = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class `synthetic_lethal`.}

\item{type}{Either `plasmid` or `treat`.}

\item{remove_multimappers}{Logical, whether to remove genes that have at least 1 guide that is a multi-mapper. Issues a warning if any of the genes appear in the new or lost essential groups.}

\item{library_name}{Name of the library, e.g. "yusa_v3_human" - must match naming conventions used in the AZ-CRUK reference data repo.}

\item{library_type}{One of "n" (knock-out), "a" (activation), or "i" (interference).}

\item{library_annotation_version}{An integer giving the annotation version of the library.}

\item{...}{Other arguments to pass to `plot`.}
}
\value{
Renders a plot in the console.
}
\description{
Plot of Treatment-vs-plasmid against Control-vs-plasmid (logFC scatterplots), and volcano plot of Treatment-vs-Control. New essential genes and lost essential genes are highlighted, but only new essential genes are annotated with gene names.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
