% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perf_curves_fgcQC.R
\name{plot_perf_curves_fgcQC}
\alias{plot_perf_curves_fgcQC}
\title{plot_perf_curves_fgcQC}
\usage{
plot_perf_curves_fgcQC(qc_fgcQC, type)
}
\arguments{
\item{qc_fgcQC}{An object of class `fgcQC`.}

\item{type}{A character string naming the plot type. One of: `ROC` or `PrRc`.}
}
\value{
A ROC or Precision-Recall curve for sets of essential genes.
}
\description{
Plots a ROC or Precision-Recall curve for sets of essential genes taking the `fgcQC` output R object as input.
}
