% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PrRc_fgcQC.R
\name{plot_PrRc_fgcQC}
\alias{plot_PrRc_fgcQC}
\title{plot_PrRc_fgcQC}
\usage{
plot_PrRc_fgcQC(data, title, print_auprrc = TRUE)
}
\arguments{
\item{data}{A data frame containing the following columns: `Recall`, `Precision`, and `gene_set`.}

\item{title}{The plot title.}

\item{print_auprrc}{Logical indicating whether a table of AUROC values should be printed to the console. Defaults to `TRUE`.}
}
\value{
A ROC plot for sets of genes.
}
\description{
Plot Precision-Recall curves for sets of essential genes.
}
