% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC_fgcQC.R
\name{plot_ROC_fgcQC}
\alias{plot_ROC_fgcQC}
\title{plot_ROC_fgcQC}
\usage{
plot_ROC_fgcQC(data, title, print_auroc = TRUE)
}
\arguments{
\item{data}{A data frame containing the following columns: `True_Positive_Rate`, `False_Positive_Rate`, and `gene_set`.}

\item{title}{The plot title.}

\item{print_auroc}{Logical indicating whether a table of AUROC values should be printed to the console. Defaults to `TRUE`.}
}
\value{
A ROC plot for sets of genes.
}
\description{
Plot ROC curves for sets of essential genes.
}
