% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plasmid_samples.R
\name{qc_plasmid_samples}
\alias{qc_plasmid_samples}
\title{qc_plasmid_samples}
\usage{
qc_plasmid_samples(data)
}
\arguments{
\item{data}{A data frame of plasmid counts, as produced by `crispRutils::read_plasmid_samples`.}
}
\value{
A data frame of summary QC metrics.
}
\description{
Calculates several QC metrics for sequenced plasmid sample counts.
}
