% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bagel_gene_results.R
\name{read_bagel_gene_results}
\alias{read_bagel_gene_results}
\title{read_bagel_gene_results}
\usage{
read_bagel_gene_results(path, ess_thresh)
}
\arguments{
\item{path}{A path to AZ-CRUK pipeline output for a single screen.}

\item{ess_thresh}{One of `orig`, `fgcQC` - indicating which essential gene sets are used for determining the Bagel threshold.}
}
\value{
An object of class `bagel_gene`.
}
\description{
Reads in all Bagel gene-level results (`*.bf`) for a single screen put through the AZ-CRUK CRISPR pipeline.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
