% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_desirability_scores.R
\name{read_desirability_scores}
\alias{read_desirability_scores}
\title{read_desirability_scores}
\usage{
read_desirability_scores(path)
}
\arguments{
\item{path}{A path to AZ-CRUK pipeline output for a single screen.}
}
\value{
A data frame.
}
\description{
Read in Desirability scores from the AZ-CRUK CRISPR analysis pipeline.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
