% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mageck_gene_results.R
\name{read_mageck_gene_results}
\alias{read_mageck_gene_results}
\title{read_mageck_gene_results}
\usage{
read_mageck_gene_results(path)
}
\arguments{
\item{path}{A path to AZ-CRUK pipeline output for a single screen.}
}
\value{
An object of class `mageck_gene`.
}
\description{
Reads in all Mageck gene-level results (`*gene_summary.txt`) for a single screen put through the AZ-CRUK CRISPR pipeline.
}
