% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plasmid_samples.R
\name{read_plasmid_samples}
\alias{read_plasmid_samples}
\title{read_plasmid_samples}
\usage{
read_plasmid_samples(dir)
}
\arguments{
\item{dir}{A character string giving a directory path containing one or more plasmid count files.}
}
\value{
A data frame.
}
\description{
Reads in multiple plasmid count samples and returns a single data frame.
}
