% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_synth_lethal_analysis.R
\name{run_synth_lethal_analysis}
\alias{run_synth_lethal_analysis}
\title{run_synth_lethal_analysis}
\usage{
run_synth_lethal_analysis(path)
}
\arguments{
\item{path}{A path to AZ-CRUK pipeline output for a single 'treatment-control' screen.}
}
\value{
An object of class `synthetic_lethal`.
}
\description{
Performs a synthetic-lethal analysis for a single screen in which a genetically modified (GM; usually a gene knock-out) isogenic cell line is compared to the unmodified parental line: the GM line is considered a 'treatment' and the parental line is the 'control' arm (as in a drug-interaction screen).
}
