% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_essentials.R
\name{union_essentials}
\alias{union_essentials}
\title{union_essentials}
\usage{
union_essentials(mageck, bagel)
}
\arguments{
\item{mageck}{An object of class `mageck_gene`.}

\item{bagel}{An object of class `bagel_gene`.}
}
\value{
An object of class `essential_union`.
}
\description{
Return the union of essentials from objects of class `mageck_gene` and `bagel_gene`.
}
