% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_desir_threshold_for_FDR.R
\name{find_desir_threshold_for_FDR}
\alias{find_desir_threshold_for_FDR}
\title{find_desir_threshold_for_FDR}
\usage{
find_desir_threshold_for_FDR(
  data,
  start_desir,
  stop_desir,
  fdr_thresh = 0.1,
  step_size = 0.05
)
}
\arguments{
\item{data}{A data frame of Desirability scores produced by \code{crispRutils::join_desirability_scores}.}

\item{start_desir}{A numeric value giving the desirability value to start at.}

\item{stop_desir}{A numeric value giving the desirability value to finish at.}

\item{fdr_thresh}{A numeric value giving the FDR threshold to aim for. Defaults to 0.1 (10\% FDR).}

\item{step_size}{A numeric value giving the step size for sliding from the start to finish Desirability thresholds. Defaults to 0.05.}
}
\value{
A data frame of FDR estimates at each Desirability threshold.
}
\description{
Finds the Desirability threshold that corresponds to a given FDR \% for \strong{shared} hits.
}
\author{
Alex T. Kalinka \email{alex.kalinka@cancer.org.uk}
}
